pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=43.0.0", "wheel"]

[project]
authors = [
  {name = "Marco Podda", email = "marco.podda@unipi.it"},
]
dependencies = [
  "hydra-core>=1.3.1",
  "hydra-colorlog>=1.2.0",
  "rich",
  "omegaconf",
  "tqdm",
  "numpy<2.0",
  "pandas",
  "pyarrow",
  "seaborn",
  "scikit-learn",
  "matplotlib",
  "ipython",
  "jupyter",
  "biopython",
  "xgboost",
]
description = "Descriptor-Free Protective Antigen Classifier"
maintainers = [
  {name = "Marco Podda", email = "marco.podda@unipi.it"},
]
name = "dfpac"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[project.urls]
"Homepage" = "https://github.com/marcopodda/dfpac.git"

[project.scripts]
dfpac-eval = "dfpac.run:main"

[tool.setuptools.packages.find]
include = ["dfpac*"] # package names should match these glob patterns (["*"] by default)
where = ["src"] # list of folders that contain the packages (["."] by default)

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
